Calibration and heat automate modules
