+++ src/server/RHData.py
@@ -14,7 +14,6 @@ import shutil
 import json
 import glob
 import RHUtils
-import random
 from six import unichr
 import Database
 import Results
@@ -1295,225 +1294,6 @@ class RHData():
                 seat += 1
 
         self.commit()
-
-    def calc_heat_pilots(self, heat_or_id):
-        heat = self.resolve_heat_from_heat_or_id(heat_or_id)
-
-        result = {
-             'calc_success': True,
-             'has_calc_pilots': False,
-             'unassigned_slots': 0
-             }
-
-        if not heat:
-            logger.error('Requested invalid heat {}'.format(heat.id))
-            result['calc_success'] = False
-            return result
-
-        # skip if heat status confirmed
-        if (heat.status == HeatStatus.CONFIRMED):
-            result['calc_success'] = None
-            logger.debug("Skipping pilot recalculation: Heat confirmed (heat {})".format(heat.id))
-            return result
-
-        # don't alter if saved races exist
-        race_list = self._Database.SavedRaceMeta.query.filter_by(heat_id=heat.id).all()
-        if (race_list):
-            result['calc_success'] = None
-            logger.debug("Skipping pilot recalculation: Races exist (heat {})".format(heat.id))
-            return result
-
-        slots = self.get_heatNodes_by_heat(heat.id)
-        for slot in slots:
-            if slot.method == ProgramMethod.NONE:
-                slot.pilot_id = RHUtils.PILOT_ID_NONE
-
-            elif slot.method == ProgramMethod.HEAT_RESULT:
-                if slot.seed_id:
-                    if slot.seed_rank:
-                        result['has_calc_pilots'] = True
-                        logger.debug('Seeding Slot {} from Heat {}'.format(slot.id, slot.seed_id))
-                        seed_heat = self.get_heat(slot.seed_id)
-
-                        if seed_heat:
-                            output = self.get_results_heat(seed_heat)
-                            if output:
-                                results = output[output['meta']['primary_leaderboard']]
-                                if slot.seed_rank - 1 < len(results):
-                                    slot.pilot_id = results[slot.seed_rank - 1]['pilot_id']
-                                else:
-                                    slot.pilot_id = RHUtils.PILOT_ID_NONE
-                                    result['unassigned_slots'] += 1
-                            else:
-                                logger.debug("Can't assign pilot from heat {}: Results not available".format(slot.seed_id))
-                                slot.pilot_id = RHUtils.PILOT_ID_NONE
-                                result['unassigned_slots'] += 1
-                        else:
-                            result['calc_success'] = False
-                            logger.info("Can't seed from heat {}: does not exist".format(slot.seed_id))
-                    else:
-                        result['calc_success'] = False
-                        logger.info("Can't seed from heat {}: rank is null".format(slot.seed_id))
-                else:
-                    slot.pilot_id = RHUtils.PILOT_ID_NONE
-                    logger.debug("Ignoring null heat as seed source")
-
-            elif slot.method == ProgramMethod.CLASS_RESULT:
-                if slot.seed_id:
-                    if slot.seed_rank:
-                        result['has_calc_pilots'] = True
-                        logger.debug('Seeding Slot {} from Class {}'.format(slot.id, slot.seed_id))
-                        seed_class = self.get_raceClass(slot.seed_id)
-
-                        if seed_class:
-                            positions = None
-
-                            ranking = self.get_ranking_raceClass(seed_class)
-                            if ranking: # manual ranking
-                                positions = ranking['ranking']
-                            else: # auto ranking
-                                results = self.get_results_raceClass(seed_class)
-                                if results:
-                                    positions = results[results['meta']['primary_leaderboard']]
-
-                            if positions:
-                                if slot.seed_rank - 1 < len(positions):
-                                    slot.pilot_id = positions[slot.seed_rank - 1]['pilot_id']
-                                else:
-                                    slot.pilot_id = RHUtils.PILOT_ID_NONE
-                                    result['unassigned_slots'] += 1
-                            else:
-                                logger.debug("Can't assign pilot from class {}: Results not available".format(slot.seed_id))
-                                slot.pilot_id = RHUtils.PILOT_ID_NONE
-                                result['unassigned_slots'] += 1
-                        else:
-                            result['calc_success'] = False
-                            logger.info("Can't seed from class {}: does not exist".format(slot.seed_id))
-                    else:
-                        result['calc_success'] = False
-                        logger.info("Can't seed from class {}: rank is null".format(slot.seed_id))
-                else:
-                    slot.pilot_id = RHUtils.PILOT_ID_NONE
-                    logger.debug("Ignoring null class as seed source")
-
-            logger.debug('Slot {} Pilot is {}'.format(slot.id, slot.pilot_id if slot.pilot_id else None))
-
-        self.commit()
-        return result
-
-    def run_auto_frequency(self, heat_or_id, current_frequencies, num_nodes, calc_fn):
-        logger.debug('running auto-frequency with {}'.format(calc_fn))
-        heat = self.resolve_heat_from_heat_or_id(heat_or_id)
-        slots = self.get_heatNodes_by_heat(heat.id)
-
-        if heat.auto_frequency:
-            # clear all node assignments
-            for slot in slots:
-                slot.node_index = None
-
-            # collect node data
-            available_seats = []
-            profile_freqs = json.loads(current_frequencies)
-            for node_index in range(num_nodes):
-                if profile_freqs["f"][node_index] != RHUtils.FREQUENCY_ID_NONE:
-                    available_seats.append({
-                        'idx': node_index,
-                        'frq': {
-                            'f': profile_freqs["f"][node_index],
-                            'b': profile_freqs["b"][node_index],
-                            'c': profile_freqs["c"][node_index]
-                            },
-                        'matches': []
-                        })
-
-            # get frequency matches from pilots
-            for slot in slots:
-                if slot.pilot_id:
-                    used_frequencies_json = self.get_pilot(slot.pilot_id).used_frequencies
-                    if used_frequencies_json:
-                        used_frequencies = json.loads(used_frequencies_json)
-                        for node in available_seats:
-                            end_idx = len(used_frequencies) - 1
-                            for f_idx, pilot_freq in enumerate(used_frequencies):
-                                if node['frq']['f'] == pilot_freq['f']:
-                                    node['matches'].append({
-                                            'slot': slot,
-                                            'band': pilot_freq['b'],
-                                            'priority': True if f_idx == end_idx else False
-                                         })
-
-            eliminated_matches = []
-            if callable(calc_fn):
-                while len(available_seats):
-                    # request assignment from calc function
-                    m_node, m_slot, an_idx = calc_fn(available_seats)
-                    if m_node and m_slot:
-                        # calc function returned assignment
-                        m_slot.node_index = m_node['idx']
-                        for slot_idx, slot_match in enumerate(m_node['matches']):
-                            if slot_match['slot'] != m_slot:
-                                eliminated_matches.append(slot_match)
-                        del available_seats[an_idx]
-                        for available_node in available_seats:
-                            for slot_idx, slot_match in enumerate(available_node['matches']):
-                                if slot_match['slot'] == m_slot:
-                                    available_node['matches'][slot_idx] = None
-                                available_node['matches'] = [x for x in available_node['matches'] if x is not None]
-                    else:
-                        # calc function didn't make an assignment
-                        random.shuffle(available_seats)
-                        if len(eliminated_matches):
-
-                            for slot_idx, slot_match in enumerate(eliminated_matches):
-                                if eliminated_matches[slot_idx]['slot'].node_index is None:
-                                    # Stay on D-band if needed
-                                    if eliminated_matches[slot_idx] \
-                                    and eliminated_matches[slot_idx]['band'] == 'D' \
-                                    and eliminated_matches[slot_idx]['priority'] == True:
-                                        for n_idx, node in enumerate(available_seats):
-                                            if node['frq']['b'] == 'D':
-                                                eliminated_matches[slot_idx]['slot'].node_index = available_seats[n_idx]['idx']
-                                                available_seats[n_idx] = None
-                                                break
-                                    else:
-                                        # else explicity avoid D-band
-                                        for n_idx, node in enumerate(available_seats):
-                                            if node['frq']['b'] != 'D':
-                                                eliminated_matches[slot_idx]['slot'].node_index = available_seats[n_idx]['idx']
-                                                available_seats[n_idx] = None
-                                                break
-
-                                    available_seats = [x for x in available_seats if x is not None]
-                                eliminated_matches[slot_idx] = None
-
-                            if len(available_seats):
-                                # can't keep D/non-D but nodes not full
-                                for slot_idx, slot_match in enumerate(eliminated_matches):
-                                    if eliminated_matches[slot_idx] and eliminated_matches[slot_idx]['slot'].node_index is None:
-                                        eliminated_matches[slot_idx]['slot'].node_index = available_seats[0]['idx']
-                                        del(available_seats[0])
-                                    eliminated_matches[slot_idx] = None
-                                
-                            eliminated_matches = [x for x in eliminated_matches if x is not None]
-                        else:
-                            # place pilots with no history into first available slots
-                            for slot in slots:
-                                if slot.node_index is None and slot.pilot_id:
-                                    if len(available_seats):
-                                        slot.node_index = available_seats[0]['idx']
-                                        del(available_seats[0])
-                                    else:
-                                        logger.warning("Dropping pilot {}; No remaining available nodes for slot {}".format(slot.pilot_id, slot))
-                            break
-            else:
-                logger.error('calc_fn is not a valid auto-frequency algortihm')
-                return False
-
-            self.commit()
-        else:
-            logger.debug('requested auto-frequency when disabled')
-
-        return True
 
     def get_results_heat(self, heat_or_id):
         heat = self.resolve_heat_from_heat_or_id(heat_or_id)
