+++ src/server/RHUtils.py
@@ -288,82 +288,6 @@ def launchBrowser(hostStr, httpPortNum=0, pageNameStr=None, launchCmdStr=None):
     except Exception:
         logger.exception("Error launching browser")
 
-# Auto-frequency algorithm prioritizing minimum channel changes
-def find_best_slot_node_basic(available_seats):
-    # if only one match has priority
-    for an_idx, node in enumerate(available_seats):
-        num_priority = 0
-        best_match = 0
-        for idx, option in enumerate(node['matches']):
-            if option['priority']:
-                num_priority += 1
-                best_match = idx
-
-        if num_priority == 1:
-            return node, node['matches'][best_match]['slot'], an_idx
-
-    # if any match has priority
-    for an_idx, node in enumerate(available_seats):
-        order = list(range(len(node['matches'])))
-        random.shuffle(order)
-        for idx in order:
-            if node['matches'][idx]['priority']:
-                return node, node['matches'][idx]['slot'], an_idx
-
-    # if only match
-    for an_idx, node in enumerate(available_seats):
-        if len(node['matches']) == 1:
-            return node, node['matches'][0]['slot'], an_idx
-
-    # if any match
-    for an_idx, node in enumerate(available_seats):
-        if len(node['matches']):
-            idx = random.randint(0, len(node['matches']) - 1)
-            return node, node['matches'][idx]['slot'], an_idx
-
-    return None, None, None
-
-# Auto-frequency algorithm suitable for Adaptive Calibration
-def find_best_slot_node_adaptive(available_seats):
-    # if only match has priority
-    for an_idx, node in enumerate(available_seats):
-        if len(node['matches']) == 1:
-            if node['matches'][0]['priority']:
-                return node, node['matches'][0]['slot'], an_idx
-
-    # if only match
-    for an_idx, node in enumerate(available_seats):
-        if len(node['matches']) == 1:
-            return node, node['matches'][0]['slot'], an_idx
-
-    # if one match has priority
-    for an_idx, node in enumerate(available_seats):
-        num_priority = 0
-        best_match = 0
-        for idx, option in enumerate(node['matches']):
-            if option['priority']:
-                num_priority += 1
-                best_match = idx
-
-        if num_priority == 1:
-            return node, node['matches'][best_match]['slot'], an_idx
-
-    # if any match has priority
-    for an_idx, node in enumerate(available_seats):
-        order = list(range(len(node['matches'])))
-        random.shuffle(order)
-        for idx in order:
-            if node['matches'][idx]['priority']:
-                return node, node['matches'][idx]['slot'], an_idx
-
-    # if any match
-    for an_idx, node in enumerate(available_seats):
-        if len(node['matches']):
-            idx = random.randint(0, len(node['matches']) - 1)
-            return node, node['matches'][idx]['slot'], an_idx
-
-    return None, None, None
-
 def getFastestSpeedStr(rhapi, spoken_flag, sel_pilot_id=None):
     fastest_str = ""
     lap_splits = rhapi.db.lap_splits()
